chore(node): move genesis to runtime
Move the Frequency and Paseo genesis configurations to the runtime.
Additionally, refactor the genesis builder for Paseo, Frequency,
and Development into a single genesis_builder function.
