+++ runtime/frequency/src/apis.rs
@@ -15,9 +15,6 @@ use sp_runtime::{
 use sp_std::prelude::Vec;
 use sp_version::RuntimeVersion;
 
-// Genesis preset configurations.
-use crate::genesis_config_presets;
-
 use super::{
 	AccountId, Balance, Block, Executive, InherentDataExt, Runtime, RuntimeCall,
 	RuntimeGenesisConfig, SessionKeys, System, TransactionPayment, VERSION,
@@ -152,12 +149,15 @@ impl_runtime_apis! {
 		}
 
 		fn get_preset(id: &Option<sp_genesis_builder::PresetId>) -> Option<Vec<u8>> {
-			get_preset::<RuntimeGenesisConfig>(id,  &genesis_config_presets::get_preset)
+			get_preset::<RuntimeGenesisConfig>(id,  &crate::genesis::presets::get_preset)
 		}
 
 		fn preset_names() -> Vec<sp_genesis_builder::PresetId> {
 			vec![
 				sp_genesis_builder::PresetId::from("development"),
+				sp_genesis_builder::PresetId::from("frequency-local"),
+				sp_genesis_builder::PresetId::from("frequency-testnet"),
+				sp_genesis_builder::PresetId::from("frequency"),
 			]
 		}
 	}
